+++ pysaliency/external_models/__init__.py
@@ -1,19 +1,18 @@
-from .matlab_models import (
-    AIM,
-    SUN,
-    ContextAwareSaliency,
-    BMS,
-    GBVS,
-    GBVSIttiKoch,
-    Judd,
-    IttiKoch,
-    RARE2012,
-    CovSal,
-)
-
 from .deepgaze import (
     DeepGazeI,
     DeepGazeIIE,
 )
-
+from .matlab_models import (
+    AIM,
+    BMS,
+    GBVS,
+    RARE2007,
+    RARE2012,
+    SUN,
+    ContextAwareSaliency,
+    CovSal,
+    GBVSIttiKoch,
+    IttiKoch,
+    Judd,
+)
 from .utils import ExternalModelMixin